backend/requirements.txt
ackermann-msgs==2.0.2
action-msgs==1.2.1
action-tutorials-interfaces==0.20.5
action-tutorials-py==0.20.5
actionlib-msgs==4.2.4
ament-cmake-test==1.3.10
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
angles==1.15.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.87.0
anyio==4.13.0
builtin-interfaces==1.2.1
certifi==2026.2.25
click==8.3.1
composition-interfaces==1.2.1
control-msgs==4.6.0
controller==2023.1.3
controller-manager==2.43.1
controller-manager-msgs==2.43.1
cv-bridge==3.2.1
demo-nodes-py==0.20.5
diagnostic-msgs==4.2.4
distro==1.9.0
docstring_parser==0.17.0
domain-coordinator==0.10.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.2
examples-rclpy-minimal-action-client==0.15.2
examples-rclpy-minimal-action-server==0.15.2
examples-rclpy-minimal-client==0.15.2
examples-rclpy-minimal-publisher==0.15.2
examples-rclpy-minimal-service==0.15.2
examples-rclpy-minimal-subscriber==0.15.2
exceptiongroup==1.3.1
fastapi==0.135.2
generate-parameter-library-py==0.3.8
geometry-msgs==4.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
image-geometry==3.2.1
interactive-markers==2.3.2
jiter==0.13.0
laser-geometry==2.4.0
launch==1.0.6
launch-ros==0.19.7
launch-testing==1.0.6
launch-testing-ros==0.19.7
launch-xml==1.0.6
launch-yaml==1.0.6
lifecycle-msgs==1.2.1
logging-demo==0.20.5
map-msgs==2.1.0
message-filters==4.3.5
mocap4r2-control-msgs==0.0.7
nav-msgs==4.2.4
osrf-pycommon==2.0.2
pcl-msgs==1.0.0
pendulum-msgs==0.20.5
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-qt-binding==1.1.2
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.5
rcl-interfaces==1.2.1
rclpy==3.3.14
rcutils==5.1.6
resource-retriever==3.1.2
rmw-dds-common==1.6.0
ros2action==0.18.11
ros2bag==0.15.12
ros2cli==0.18.11
ros2component==0.18.11
ros2doctor==0.18.11
ros2interface==0.18.11
ros2launch==0.19.7
ros2lifecycle==0.18.11
ros2multicast==0.18.11
ros2node==0.18.11
ros2param==0.18.11
ros2pkg==0.18.11
ros2run==0.18.11
ros2service==0.18.11
ros2topic==0.18.11
rosbag2-interfaces==0.15.12
rosbag2-py==0.15.12
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpyutils==0.2.1
rqt==1.1.7
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-controller-manager==2.43.1
rqt-gauges==0.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-joint-trajectory-controller==2.37.3
rqt-moveit==1.0.1
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==1.0.6
rqt-robot-steering==1.0.0
rqt-runtime-monitor==1.0.0
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-tf-tree==1.0.5
rqt-topic==1.5.0
sensor-msgs==4.2.4
sensor-msgs-py==4.2.4
shape-msgs==4.2.4
sniffio==1.3.1
sros2==0.10.5
starlette==1.0.0
statistics-msgs==1.2.1
std-msgs==4.2.4
std-srvs==4.2.4
stereo-msgs==4.2.4
teleop-twist-keyboard==2.4.0
tf2-geometry-msgs==0.25.8
tf2-kdl==0.25.8
tf2-msgs==0.25.8
tf2-py==0.25.8
tf2-ros-py==0.25.8
tf2-tools==0.25.8
topic-monitor==0.20.5
trajectory-msgs==4.2.4
turtlesim==1.4.2
typing-inspection==0.4.2
typing_extensions==4.15.0
unique-identifier-msgs==2.2.1
uvicorn==0.42.0
vehicle==2023.1.3
vision-msgs==4.1.1
visualization-msgs==4.2.4
webots-ros2==2023.1.3
webots-ros2-driver==2023.1.3
webots-ros2-epuck==2023.1.3
webots-ros2-importer==2023.1.3
webots-ros2-mavic==2023.1.3
webots-ros2-msgs==2023.1.3
webots-ros2-tesla==2023.1.3
webots-ros2-tiago==2023.1.3
webots-ros2-turtlebot==2023.1.3
webots-ros2-universal-robot==2023.1.3
xacro==2.0.8
